+++ src/dk/aau/cs/ds306e18/tournament/utility/AutoNaming.java
@@ -32,7 +32,7 @@ public class AutoNaming {
                 .filter(t -> t != team)
                 .map(Team::getTeamName)
                 .collect(Collectors.toSet());
-        teamName = uniqueify(teamName, otherTeamNames);
+        teamName = uniquify(teamName, otherTeamNames);
         team.setTeamName(teamName);
     }
 
@@ -41,7 +41,7 @@ public class AutoNaming {
      * be added the end of the team name. E.g. "Team" already exist, then "Team" becomes "Team (1)".
      * The new unique name is returned.
      */
-    public static String uniqueify(String teamName, Set<String> otherTeamsNames) {
+    public static String uniquify(String teamName, Set<String> otherTeamsNames) {
         int i = 1;
         String candidateName = teamName;
         while (otherTeamsNames.contains(candidateName)) {
@@ -65,6 +65,7 @@ public class AutoNaming {
     /**
      * Finds a team name for the given list of bots. The algorithm will try to find an
      * interesting part from each bot's name and use it in the team name.
+     * If there is only one unique name, that name will be used.
      * The algorithm isn't perfect, and characters other than letters and digits
      * can create weird names.
      * Example: ReliefBot + Beast from the East => Relief-Beast.
@@ -73,6 +74,10 @@ public class AutoNaming {
         int botCount = botNames.size();
         if (botCount == 0) return "Team";
         if (botCount == 1) return botNames.get(0);
+
+        // Check if there is only one unique name, in that case use that name
+        long uniqueNames = botNames.stream().distinct().count();
+        if (uniqueNames == 1) return botNames.get(0);
 
         // Construct names from the short names of bots, separated by "-"
         // Example: ReliefBot + Beast from the East => Relief-Beast